Code cleanup, addressing compiler warnings upto gcc8...
    - (Re)enstate Cpu, Mem and Video information for modern platform kernels.
    - Updated internal documetation to reflect current build process.

+++ src/video/SDL_Console.cpp
@@ -180,7 +180,7 @@ void UpdateConsole()
 void DrawConsole()
 {
 	int loop;
-	SDL_Rect DestRect = {0, 0, ConsoleSurface->w, ConsoleSurface->h};
+	SDL_Rect DestRect = {0, 0, (short unsigned int)ConsoleSurface->w, (short unsigned int)ConsoleSurface->h};
 	
 
 	SDL_BlitSurface(ConsoleSurface, NULL, OutputScreen, &DestRect);
